refactor(service-log): drop redundant entity name from use-case filenames
add-service-log.ts/edit-service-log.ts/remove-service-log.ts repeated
the module name their own directory already scopes. Matches car's
use-case convention (add.ts, edit.ts, remove.ts) and service-log's own
interface/api and tanstack/mutation-options directories, which already
use bare verbs for these same three operations. Class names are
unchanged.
