+++ src/module/car/service-log/application/use-case/remove.ts
